Revert "Refactor example app structure and implement navigation; create HomePage, SingleInstancePage, and MultipleInstancesPage. Consolidate locale handling and improve NodeEditorInstance functionality. Remove unused imports and streamline clipboard and rendering utilities."
This reverts commit 0972177151b76079d5a78704bc1768b10dc5359c.
